# target/generated-sources/openapi/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:8080
...
paths:
  /locations/{pointId}:
    delete:
      description: удаляет локацию по ID
      operationId: deleteLocation
      parameters:
      - explode: false
        in: path
        name: pointId
        required: true
        schema:
          format: int64
          minimum: 1
        style: simple
      responses:
        "400":
          description: Невалидные параметры запроса
        "401":
          description: Неверные авторизационные данные
        "200":
          description: Локация успешно удалена
        "404":
          description: Точка локации с таким ID не найдена
      summary: Удалить локацию
      tags:
      - Локации
      x-accepts: application/json
      x-tags:
      - tag: Локации
    put:
      description: Обновляет данные локации и возвращает обновленные данные
      operationId: updateLocation
      parameters:
      - explode: false
        in: path
        name: pointId
        required: true
        schema:
          format: int64
          minimum: 1
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationDtoRequest'
        required: true
      responses:
        "409":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationDtoResponse'
          description: Точка локации с такими latitude и longitude уже сущесвтует
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationDtoResponse'
          description: Локация успешно обновлена
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationDtoResponse'
          description: Невалидные параметры запроса
        "404":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationDtoResponse'
          description: Точка локации с таким ID не найдена
      summary: Обновить данные локации
      tags:
      - Локации
      x-content-type: application/json
      x-accepts: '*/*'
      x-tags:
      - tag: Локации
...
  /registration:
    post:
      description: Регистрирует новых пользователй
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDtoRequest'
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountDtoResponse'
          description: Запрос усешно выполнен
        "403":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountDtoResponse'
          description: Запрос от авторизованного аккаунта
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountDtoResponse'
          description: Невалидные данные для регистрации
        "409":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountDtoResponse'
          description: Аккаунт с таким email уже существует
      summary: Зарегистрироваться
      tags:
      - Registration
      x-content-type: application/json
      x-accepts: '*/*'
      x-tags:
      - tag: Registration
  /locations:
    post:
      description: Создает новую локацию по введенным параметрам и возвращает добавленную
        локацию
      operationId: addLocation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationDtoRequest'
        required: true
      responses:
        "409":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationDtoResponse'
          description: Точка локации с такими latitude и longitude уже сущесвтует
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationDtoResponse'
          description: Невалидные параметры запроса
        "401":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationDtoResponse'
          description: Неверные авторизационные данные
        "201":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationDtoResponse'
          description: Локация успешно создана
      summary: Добавить локацию
      tags:
      - Локации
      x-content-type: application/json
      x-accepts: '*/*'
      x-tags:
      - tag: Локации
...
  /locations/{id}:
    get:
      description: Возвращает локацию по ID
      operationId: getLocation
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          minimum: 1
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationDtoResponse'
          description: Запрос успешно выполнен
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationDtoResponse'
          description: Невалидные параметры запроса
        "401":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationDtoResponse'
          description: Неверные авторизационные данные
        "404":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationDtoResponse'
          description: Точка локации с таким ID не найдена
      summary: Получение локации
      tags:
      - Локации
      x-accepts: '*/*'
      x-tags:
      - tag: Локации
...
components:
  schemas:
    LocationDtoRequest:
      example:
        latitude: ""
        longitude: ""
      properties:
        latitude:
          format: double
          maximum: 90
          minimum: -90
        longitude:
          format: double
          maximum: 180
          minimum: -180
      required:
      - latitude
      - longitude
    LocationDtoResponse:
      example:
        latitude: ""
        id: ""
        longitude: ""
      properties:
        id:
          format: int64
        latitude:
          format: double
        longitude:
          format: double
...
    AccountDtoRequest:
      example:
        firstName: ""
        lastName: ""
        password: ""
        email: ""
      properties:
        firstName: {}
        lastName: {}
        email: {}
        password: {}
      required:
      - email
      - firstName
      - lastName
      - password
    AccountDtoResponse:
      example:
        firstName: ""
        lastName: ""
        id: ""
        email: ""
      properties:
        id:
          format: int32
        firstName: {}
        lastName: {}
        email: {}
...